Adding AskQuestions function Refactoring biceps

+++ deploy.ps1
@@ -45,28 +45,31 @@ Write-Host "Getting current user object id" -ForegroundColor DarkCyan
 $currentUserObjectId = az ad signed-in-user show -o tsv --query id
 Write-Host "Current User Object Id: $currentUserObjectId" -ForegroundColor Green
 
-if($loadBalancingType -eq "priority")
-{
-	Write-Host "Building priority load balancing policy" -ForegroundColor DarkCyan
-	$arrayList = [System.Collections.ArrayList]::new()
-	$params = (az bicep build-params --file ./infra/main.bicepparam --stdout | ConvertFrom-Json -depth 100 ).parametersJson | ConvertFrom-Json -depth 100
-	foreach($oai in  $params.parameters.openAIInstances.value)
-	{
-		$beName = "OPENAI$($oai.suffix.ToUpper())"
-		$item = "backends.Add(new JObject() {{ ""backend-id"", ""$($beName)"" }, { ""priority"", $($oai.priority) }, { ""isThrottling"", false }, { ""retryAfter"", DateTime.MinValue }}); "
-		$arrayList.Add($item)
-	}
-    
-	$policy = Get-Content "./infra/APIM/priority-load-balance-policy.xml" -Raw
-	$policy = $policy -replace "{{BACKENDS}}", ($arrayList -join "`r`n`t`t`t`t`t")
-	Write-Host "Writing priority load balancing policy XML with $($arrayList.Count) backends" -ForegroundColor DarkCyan
-    $policy | Set-Content -Path "./infra/APIM/priority-load-balance-policy-main.xml"
-}
+
 
 if(!$?){ exit }
 
 if($codeDeployOnly -eq $false)
 {
+
+	if($loadBalancingType -eq "priority")
+	{
+		Write-Host "Building priority load balancing policy" -ForegroundColor DarkCyan
+		$arrayList = [System.Collections.ArrayList]::new()
+		$params = (az bicep build-params --file ./infra/main.bicepparam --stdout | ConvertFrom-Json -depth 100 ).parametersJson | ConvertFrom-Json -depth 100
+		foreach($oai in  $params.parameters.openAIInstances.value)
+		{
+			$beName = "OPENAI$($oai.suffix.ToUpper())"
+			$item = "backends.Add(new JObject() {{ ""backend-id"", ""$($beName)"" }, { ""priority"", $($oai.priority) }, { ""isThrottling"", false }, { ""retryAfter"", DateTime.MinValue }}); "
+			$arrayList.Add($item)
+		}
+		
+		$policy = Get-Content "./infra/APIM/priority-load-balance-policy.xml" -Raw
+		$policy = $policy -replace "{{BACKENDS}}", ($arrayList -join "`r`n`t`t`t`t`t")
+		Write-Host "Writing priority load balancing policy XML with $($arrayList.Count) backends" -ForegroundColor DarkCyan
+		$policy | Set-Content -Path "./infra/APIM/priority-load-balance-policy-main.xml"
+	}
+
 	Write-Host "Deploying resources to Azure" -ForegroundColor DarkCyan
 	$output = az deployment sub create --location $location  --template-file ./infra/main.bicep `
 		--parameters ./infra/main.bicepparam `
@@ -93,6 +96,7 @@ if($codeDeployOnly -eq $false)
 	$funcAiSearch = $outputObj.properties.outputs.aiSearchIndexFunctionName.value
 	$funcMove = $outputObj.properties.outputs.moveFunctionName.value
 	$funcQueue = $outputObj.properties.outputs.queueFunctionName.value
+	$funcQuestions = $outputObj.properties.outputs.questionsFunctionName.value
 
 	if(!$?){ exit }
 }
@@ -103,6 +107,7 @@ else {
 	$funcMove = "func-$appName-Mover-$location"
 	$funcQueue = "func-$appName-Queueing-$location"
 	$funcAiSearch = "func-$appName-AiSearch-$location"
+	$funcQuestions = "func-$appName-AskQuestions-$location"
 }
 if(!$?){ exit }
 
@@ -121,13 +126,26 @@ $scriptDir = Split-Path $script:MyInvocation.MyCommand.Path
 #dotnet clean -c release
 #dotnet clean -c debug
 $childPath = "bin/Release/net8.0/publish"
+$zip = $scriptDir + "build.zip"
+
+
+Write-Host "Deploying Ask Questions Function App" -ForegroundColor DarkCyan
+Push-Location -Path DocumentQuestionsFunction
+#dotnet clean .
+dotnet publish .
+$source = Join-Path -Path $pwd.Path -ChildPath $childPath
+if(Test-Path $zip) { Remove-Item $zip }
+[io.compression.zipfile]::CreateFromDirectory($source,$zip)
+az webapp deploy --name $funcQuestions --resource-group $resourceGroupName --src-path $zip --type zip
+Pop-Location
+
+if(!$?){ exit }
 
 Write-Host "Deploying Document Intelligence Function App" -ForegroundColor DarkCyan
 Push-Location -Path DocumentIntelligenceFunction
 #dotnet clean .
 dotnet publish .
 $source = Join-Path -Path $pwd.Path -ChildPath $childPath
-$zip = $scriptDir + "build.zip"
 if(Test-Path $zip) { Remove-Item $zip }
 [io.compression.zipfile]::CreateFromDirectory($source,$zip)
 az webapp deploy --name $funcProcess --resource-group $resourceGroupName --src-path $zip --type zip
@@ -169,3 +187,5 @@ if(Test-Path $zip) { Remove-Item $zip }
 [io.compression.zipfile]::CreateFromDirectory($source,$zip)
 az webapp deploy --name $funcQueue --resource-group $resourceGroupName --src-path $zip --type zip
 Pop-Location
+
+
